requirements.txt
numpy
pandas
PyYaml
scipy.signal
argparse
os
logging
sys
warnings
typing
tifex-py
scipy==1.11.4